minor change to configure.ac and Makefile.am
Makefile.am: change gnu11 to c99. That can be changed later if needed

configure.ac: use "foreign" for automake options

>Automake will check for only those things that are absolutely required
for proper operations. For instance, whereas GNU standards dictate the
existence of a NEWS file, it will not be required in this mode. This
strictness will also turn off some warnings by default (among them,
portability warnings). The name comes from the fact that Automake is
intended to be used for GNU programs; these relaxed rules are not the
standard mode of operation.

The rest of the files in this commit were auto-generated when I ran
autoreconf -ivf